Adding the entire project to the feature/map-location branch, with all proposed changes, waiting for approval
