+++ NodeServer/frontend/UI.html
@@ -1,0 +1,22 @@
+﻿<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Weather</title>
+    <link rel="stylesheet" href="css/style.css">
+    <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
+</head>
+<body class="background">
+    <div class="content">
+        <h1>Check the weather</h1>
+        <form id="weatherForm">
+            <input type="text" id="city" placeholder="Enter city name">
+            <button type="submit">Check Weather</button>
+        </form>
+        <div id="weatherResult"></div>
+        <canvas id="temperatureChart"></canvas>
+    </div>
+    <script src="js/Script.js"></script>
+</body>
+</html>
